xlab_project/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
distro==1.9.0
Django==5.0.7
djangorestframework==3.15.2
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
openai==1.36.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyOpenSSL==24.1.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
zope.interface==6.4.post2
